+++ Specs/DataSources/BillboardVisualizerSpec.js
@@ -15,8 +15,7 @@ defineSuite([
         'Scene/BillboardCollection',
         'Scene/HorizontalOrigin',
         'Scene/VerticalOrigin',
-        'Specs/createScene',
-        'Specs/destroyScene'
+        'Specs/createScene'
     ], function(
         BillboardVisualizer,
         BoundingRectangle,
@@ -33,8 +32,7 @@ defineSuite([
         BillboardCollection,
         HorizontalOrigin,
         VerticalOrigin,
-        createScene,
-        destroyScene) {
+        createScene) {
     "use strict";
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
 
@@ -46,7 +44,7 @@ defineSuite([
     });
 
     afterAll(function() {
-        destroyScene(scene);
+        scene.destroyForSpecs();
     });
 
     afterEach(function() {