+++ Specs/Scene/GlobeSurfaceTileSpec.js
@@ -19,7 +19,6 @@ defineSuite([
         'Scene/TerrainState',
         'Scene/TileImagery',
         'Specs/createContext',
-        'Specs/destroyContext',
         'ThirdParty/when'
     ], function(
         GlobeSurfaceTile,
@@ -41,7 +40,6 @@ defineSuite([
         TerrainState,
         TileImagery,
         createContext,
-        destroyContext,
         when) {
     "use strict";
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
@@ -94,7 +92,7 @@ defineSuite([
         });
 
         afterAll(function() {
-            destroyContext(context);
+            context.destroyForSpecs();
         });
 
         beforeEach(function() {
@@ -542,7 +540,7 @@ defineSuite([
         });
 
         afterAll(function() {
-            destroyContext(context);
+            context.destroyForSpecs();
         });
 
         it('gets correct results even when the mesh includes normals', function() {