+++ Specs/Widgets/CesiumInspector/CesiumInspectorSpec.js
@@ -3,14 +3,12 @@ defineSuite([
         'Widgets/CesiumInspector/CesiumInspector',
         'Core/Ellipsoid',
         'Scene/Globe',
-        'Specs/createScene',
-        'Specs/destroyScene'
+        'Specs/createScene'
     ], function(
         CesiumInspector,
         Ellipsoid,
         Globe,
-        createScene,
-        destroyScene) {
+        createScene) {
     "use strict";
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
 
@@ -23,7 +21,7 @@ defineSuite([
     });
 
     afterAll(function() {
-        destroyScene(scene);
+        scene.destroyForSpecs();
     });
 
     it('can create and destroy', function() {