+++ Specs/Widgets/HomeButton/HomeButtonSpec.js
@@ -1,12 +1,10 @@
 /*global defineSuite*/
 defineSuite([
         'Widgets/HomeButton/HomeButton',
-        'Specs/createScene',
-        'Specs/destroyScene'
+        'Specs/createScene'
     ], function(
         HomeButton,
-        createScene,
-        destroyScene) {
+        createScene) {
     "use strict";
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
 
@@ -16,7 +14,7 @@ defineSuite([
     });
 
     afterAll(function() {
-        destroyScene(scene);
+        scene.destroyForSpecs();
     });
 
     it('constructor sets default values', function() {